Added hot springs, fissures, attached gemstone veins.
- Minor other ore vein reworks, mostly of the gemstone variety.
- Hot springs (and attached fissures, decoration), Empty ones, Lava fissures on volcanoes, with decoration.
- Added the `locatevein` subcommand
- Fixed the forest groundcover using unsafe fluid replacement checks
- Added (for some reason, broken) vein biome checks for lake and volcanic veins. Codec is not parsing out them for some reason.

+++ src/main/java/net/dries007/tfc/common/command/TimeCommand.java
@@ -62,7 +62,7 @@ public final class TimeCommand
                 )
                 .then(Commands.literal("days")
                     .then(Commands.argument("value", IntegerArgumentType.integer(1))
-                        .executes(context -> addTime(IntegerArgumentType.getInteger(context, "value") * ICalendar.TICKS_IN_DAY))
+                        .executes(context -> addTime(IntegerArgumentType.getInteger(context, "value") * (long) ICalendar.TICKS_IN_DAY))
                     )
                 )
                 .then(Commands.literal("ticks")
